fix(rocky): added 2 urls parse
